FPDF_GetTrailerEnds: make this not depend on whitespace
PDF-1.7 calls out no bytes other than whitespace when specifying what
can occur between endstream and endobj, so whitespace needs to be
handled. When CPDF_SyntaxParser::ReadStream() reads the stream, then it
reads 'endobj', and then resets the position back to the end of
'endstream'. This mechanism is disabled in case there is whitespace
between the tokens and the newline, see the end of the function.

This results in reporting no trailer ends, as the parsing fails, as the
next token is expected to be 'endobj', but it's the ID of the next
object instead.

Fix the problem by handling whitespace in
CPDF_SyntaxParser::ReadStream() where it was looking for \ntoken\n, not
allowing whitespace between the token and the following newline.

Change-Id: I7048e8d081af04af3dd08d957212c885b7982b5e