Rename CPDF_ReadValidator::Session to ScopedSession
Whilst reading through this code, it took me 30 seconds to realize
that the reason for these session objects was solely to restore state
when they went out of scope, so name them as such to make this clearer
in the future.

Make members const, then remove const from callers, which also aided
in my reading of the code.

Change-Id: I00952e19d9c173072bc0b89a824085c02a2ba9c5