Rename CPDF_ReadValidator::Session to ScopedSession
Whilst reading through this code, it took me 30 seconds to realize
that the reason for these session objects was solely to restore state
when they went out of scope, so name them as such to make this clearer
to future readers.

Make members const, then remove const from callers, which also aided
in my reading of the code, since actions happen on cleanup which,
although legal, don't feel like the spirit of const.

Mark the class as stack-allocated only, then install a RetainPtr<>
rather than an UnownedPtr<> to the validator, since there can't
be any cycles and it can't hurt to hold that additional reference.

-- Initialize CPDF_ReadValidator members in header while at it.
-- Mark some CPDF_ReadValidator members const while at it.

Change-Id: I00952e19d9c173072bc0b89a824085c02a2ba9c5