Enable PDFFlattenEmbedderTest.BUG_889099 for Skia/SkiaPaths.
When Skia/SkiaPaths is enabled, FPDFFlattenEmbedderTest.BUG_889099 have
small discrepancies on text rendering compared to AGG's expected results,
which are caused by Skia using different anti-aliasing and font face
rendering methods compared to AGG. Also there is small difference on the
background color for the matching results for |kPageHash|, which is
caused by loss of precision during Skia's premultiplied process (see
https://crbug.com/pdfium/1515#c1). However Skia and SkiaPaths still
generate acceptable rendering results.

This CL updates the Skia/SkiaPaths expected results for the test so
that it can be enabled and pass.

Bug: pdfium: 1500
Change-Id: Ifcd02abdf183e44dce8828c267a2d9d83b8e1e79