Make cppgc initialization the responsiblity of the embedder.
This is a breaking change for XFA builds.

This CL keeps PDFium up-to-date with the new method of
initializing the cppgc ("oilpan") garbage collector used
by XFA. It is now cumbersome to do this as part of library
initialization, so much like v8 itself, the embedder must
set up the cppgc environment itself.

The upshot is that embedders using the XFA feature must
search for their existing calls to v8::InitializePlatform()
and follow them up by a call to cppgc::InitializeProcess().

Bug: chromium:1176416
Change-Id: Ib3c39cd7933c8f9cf048840bd9206672434c7159