Make cppgc initialization the responsibility of the embedder.
*** This is a breaking change for XFA builds. ***

This CL keeps PDFium up-to-date with the new method of initializing the
cppgc ("oilpan") garbage collector used by XFA following the change at
https://chromium.googlesource.com/v8/v8/+/8c99b253af975acf09d8239e1798484d5a8a82fe

It is now cumbersome for PDFium to do this as part of its library
initialization, so much like chromium, the embedder must set up the
cppgc environment itself.

The upshot is that embedders using the XFA feature must search for their
existing calls to v8::InitializePlatform() and follow them up by a call
to cppgc::InitializeProcess().

-- Introduce new function in v8_initializer.cpp to keep cleanup in
   the same file as initialization.

Bug: chromium:1176416
Change-Id: Ib3c39cd7933c8f9cf048840bd9206672434c7159