Move GetZeroAreaPath() to core/fxge/cfx_renderdevice.cpp.
In CFX_RenderDevice::DrawPathWithBlend(), to process each sub path from
`pPathData`, we will need to form every sub path in that same function
and use GetZeroAreaPath() to calculate its zero areas before drawing.
This requires GetZeroAreaPath() taking any vector of FX_PATHPOINT as a
path and processing it, instead of processing all the points of
`pPathData`. Therefore GetZeroAreaPath() no longer needs to be a member
of CFX_PathDataon.

This CL makes GetZeroAreaPath() take any vector of FX_PATHPOINTs as an
input parameter, and moves it and its helper functions to the anonymous
namespace in core/fxge/cfx_renderdevice.cpp.

Bug: pdfium:1638
Change-Id: Ia07cbb261a09bd002ed56b976d450ed2951c9cc9