Split a path into sub paths for drawing zero area.
Currently GetZeroAreaPath() is used to calculate the zero areas formed
by a whole path. However, its output parameter `thin` `set_identity` can
be set due to a sub path (a subset of the whole path that started with
"move to" action), while affecting the rendering results of all the zero
areas formed by the whole path.

This CL adds the process to split the whole path into sub paths, and
calls DrawZeroAreaPath() to process each sub path individually.

This CL also fix an issue that horizontal and vertical folding lines
are not rendered thin and uploads a pixel test for it.

Bug: pdfium:1638
Change-Id: I5492870e5c7e9254de346cc7d7c9255661e3abd2