Enable FPDFEditEmbedderTest.SaveAndRender for SkiaPaths.
For FPDFEditEmbedderTest.SaveAndRender, the rendering results have
been uploaded at [1]. There are barely visible discrepancies between
the AGG and SkiaPaths results which are casued by different
anti-aliasing and Bezier curve algorithms, and the rendering results
are both acceptable. However, the rendering result of Skia shows that
the stroke color and the fill color overlapping along the triangle
path, which is not ideal.

Therefore only enable this embedder test for SkiaPaths and link the
Skia issue with [2].

[1] https://bugs.chromium.org/p/pdfium/issues/detail?id=1500#c96
[2] https://crbug.com/pdfium/1651

Bug: pdfium:1500
Change-Id: I1a124565f2de07dfb60dff79b5cce9ecb267c0fe