Avoid uniuqe_ptrs to string streams in CPDF_PageContentGenerator.
Move-assignment of a string stream should be a no-copy operation
and reasonably efficient, so the additional indirection through
the unique_ptr isn't needed.

The only compication is we hit a "const" issue in one place, so
convert one method to take an rvalue reference to avoid passing
args by non-const ref.

Change-Id: Ie7b8aa1a2b1a864521faa30ebcb2343861064620