Add a test case for FPDFAnnot_SetBorder() to test rendering.
The existing test case that exercises FPDFAnnot_SetBorder() cannot
render the line annotations in its test PDF. Rather than waiting for
that issue to get resolved, just leave the existing test case alone and
write a second test that uses a different annotation type which is known
to have support for borders.

Add a new square annotation to annots.pdf, and use that in a new test
case to make sure FPDFAnnot_SetBorder() calls reflect in the rendered
in-memory PDF and in the saved PDF. With only the new test, it does not
work as expected and shows FPDFAnnot_SetBorder() does not work properly.
Change FPDFAnnot_SetBorder() to remove appearance streams to resolve the
issue and make the test pass.

Change-Id: I0f34832c545f8d29350542bb57ab1de5f14b39f4