Reduce no_chromium_code build config usage.
Various build targets, including some in third_party/, are really first
party code in the sense that they are written by PDFium or Chromium
authors. As such, the code quality for those targets is high enough that
they can be use the "chromium_code" build config, which enables more
warnings.

Change-Id: If73f5bd4217be2f8656e2da664b59b85d7d1aa09