Rename/Re-order some CFX_DIBBase methods.
Do some tidying before attempting to encapsuate m_pAlphaMask. Doing so
will require a new HasAlphaMask() method, which sounds too much like the
existing HasAlpha() method, so rename that to IsAlphaFormat(). Then
rename IsMask() to IsMaskFormat() for consistency. Then separate out
the pallete functions into their own section. In doing so, notice that
GetPaletteSize() is returing a number based on format, not what was
actually allocated for the palette, so rename to GetRequiredPaletteSize(),
and flag places where we are using this size rather than the actual
size with which a palette was previously allocated. Will address those
in a follow-up since should the format change, there's the possibility
of getting a mismatch.

Change-Id: I98c7fb1669a72c8d9dfd5e550b1c1bbbb9c22cb6