Rename/Re-order some CFX_DIBBase methods.
Do some tidying before attempting to encapsuate m_pAlphaMask. Straight
rename/shuffle, no funny business.

Encapsulation will require a new HasAlphaMask() method, which sounds
too much like the existing HasAlpha() method, so rename that to
IsAlphaFormat(). Then rename IsMask() to IsMaskFormat() for consistency.

Then separate out the palette functions into their own section. In
doing so, notice that GetPaletteSize() is returning a size based on
format, not what was actually allocated for the palette, so rename to
GetRequiredPaletteSize(). Then flag places where we are using this size
rather than the actual size with which a palette span was previously
allocated. Will address those in a follow-up to avoid the possibility
of ever getting a mismatch.

Change-Id: I98c7fb1669a72c8d9dfd5e550b1c1bbbb9c22cb6