Improve check to filter out obvious non-rectangle path points.
In CFX_AggDeviceDriver::SetClip_PathFill() and
CFX_RenderDevice::DrawPathWithBlend(), the rendering code that handles
rectangles will first do a simple check before calling
CFX_PathData::GetRect(), which is more expensive. The simple check is
currently part of the old CFX_PathData::GetRect() implementation, which
incorrectly assumes a set of path points must have 4 or 5 points to form
a rectangle. The correct check is 4 or more.

With this and related CLs that improved CFX_PathData, the slow rendering
PDF that triggered this investigation now renders 77 times faster
according to safetynet_measure.py with callgrind.

Bug: pdfium:1683
Change-Id: I26b75e0e72322ce409bbb15bc6b06c94660107d7