Reimplement CFX_PathData::IsRect() and GetRect().
Consolidate many duplicate checks together and make the code easier to
understand. This also fixes a disagreement where IsRect() says 4 points
do not draw a rectangle, but GetRect() says yes. Add a
GetNormalizedPoints() helper to normalize 5 point rectangles into 4
point rectangles. This helper also structures the code so it is easier
to normalize arbitrary points that form a rectangle in the (near)
future.

Bug: pdfium:1683
Change-Id: Ie70687c12dc5f7b94f1c956571047832244916c4