Remove usage of v8::Local<> in xfa_basic_data.h.
A split between the (element, attribute) table and the callback
table was introduced as part of the effort to reduce .relro sections.
As a result of this split, we can now confine any knowledge of

v8::Local<> to files in fxjs by removing the the last usage in
xfa_basic_data.h by putting the callback table into fxjs.
Change-Id: I13daaf1c65dbc558f6906000d87924648a822dd1