[code health] Avoid use of Optional<T>::operator bool()
A recent offline discussion indicates that this may be error-prone
compared to using .has_value(). Similarly, avoid operator *() in
favor of .value() for symmetry and because the precedence of `*`
isn't always what we'd like for these kinds of expressions.

Stop short of completely removing this from pdfium::Optional,
because some gtest macros require this under the cover for
our tests as written.

Change-Id: Iabb6257e53bb333d5b96d42f1f40a39d1a3da59a