Remove some unreachable code from progressive_decoder.cpp
ProgressiveDecoder::WeightTable::Calc() runs the risk of passing a
negative lenght parameter to resize() if the dest_len parameter is
less than 0. Fortunately, the only caller passes m_sizeX, which is
only set to a non-zero value from device_rc.Width() at line 2172.
We know this is non-negative since we checked size_x at line 2158
priot to setting up device_rc.

Convert to a check for non-negative values, and remove the
calculations involving `base` since we know that it will be zero.
Note that Calc() is uncomfortably simliar to code in CStretchEngine,
but that does not have the same issue because the size is computed
from different parameters.

Change-Id: I7cc3d8672036150691f93a55e05659dc8168e3f6