CStretchEngine: Safely avoid PixelWeight bounds checks
CHECK() that we are always setting the start and end pixel range
within the allocated weight array. Then replace an expensive function
with an "unsafe" one, noting that it is only used within fixed-bounds
for-loops based on the same start and end value we would check in a
"safe" version.

Change-Id: Ibffc53b3e1c3706414a01b6e5e8f638f16e70643