CStretchEngine: ensure sum of pixel weights is kFixedPointOne.
When calculating pixel weights using floating point arithemetic and
rounding, there is no guarantee that the sum of the rounding errors
will be zero. Instead, acumulate the error and include those few
fractional bits into the last weight. This may help the off-by-one
pixel values seen in the referenced bug, but may not completely
solve it.

Bug: pdfium:1688
Change-Id: I7585731ad6aca1557b46e3a26eb3ddcb0aab59ca