Expose DashingArray and Phase in the public API
Access to dashing info was available in the original Foxit api, but
not publicly accessible in pdfium. This provides getters and setters
for them in fpdf_edit.h. Additionally, CFX_GraphState was extended to
enable access. In contrast with the Foxit API, the dashing array getter
copies to an array directly instead of passing in an index, like in
FPDFPath_GetPathSegment, to reduce the need for copying.

R=tsepez@chromium.org

Bug: pdfium:1187
Change-Id: If4b376e164dc8c8a97268ff380683b5d6a9f4be4