Remove windows.h from public/fpdfview.h
This is a BREAKING CHANGE and may require embedders on windows platforms
to include this file themselves, however we believe the public APIs to
be free of windowisms and can avoid silently dragging this file into
all who embed PDFium. The result is a less polluted namespace as
windows.h introduces a number of #defines that can get in the way.

Change-Id: I480e843e2a302bd075b0b2a3f648b4eb5e7f9286