Re-introduce FWL_STATE_PSB_Default definition.
Once upon a time, there was a
  #define FWL_STATE_PSB_Default (1 << (FWL_WGTSTATE_MAX + 2))

that we removed because the definition was never used. However,
it turns out there were places where the equivalent expression
  (1 << (FWL_WGTSTATE_MAX + 2))

was used in conditionals, so restore the value, and use it there.

On the other hand, there doesn't appear to be a way to actually
set this value, but since the style extension field is overloaded
in all sorts of strange ways, we can be 100% sure that this is
dead code as a result.

Change-Id: Iab2158cfdc672a11b50bcd9b0d7ae5d7f47d8af7