Convert OnChar() arguments back to uint32_t
The CL at https://pdfium-review.googlesource.com/c/pdfium/+/83870
was a bit over-zealous in the conversion of the OnChar() methods to
VKEYCODE. Although we frequently pass VKEYCODES to these methods,
and compare against VKEYCODES inside these methods, the arguments
are actually character codes, not key codes. The cross-type
usages happen to work because we selectively use the only the
values where they are equivalent.

We'll need a subsequent CL to split CFWL_MessageKey into, say,
CFWL_MessageChar to get complete type correctness, but just
cast for now.

Change-Id: I5ada6ea2abe124116e3eaa13460cf05e2eb055d8