Zero out-params in CPDF_SeparationCS::GetRGB().
Bug was introduced in
  https://pdfium-review.googlesource.com/c/pdfium/+/2991

when we converted some out-paramters from refernces to pointers. The
current code nulls the pointers rather than zeroing the pointed-to
RGB values. Not all callers check the return value, but we do return
false without zeroing in other blocks, so the consequences are likely
to be minimal.

Change-Id: Ia6d29328af1ec65ed4b523b60bf8b8190b70198b