Rename CFLL_RadioButton::GetRadioButton() to GetPWLRadioButton()
Straight rename, no funny business.

I spent a few seconds wondering why we had to call GetRadioButton() on
a radio button and whether is was superfluous and should be removed.
It's not, since it returns an object from the lower PWL layer. The
new naming should make this more obvious.

-- Do the same for CFFL_CheckBox, CFFL_ListBox, and CFFL_TextField.

Change-Id: I908cdbfc0bca1d4c350987b0f8b7069e6d898dfe