Handle null annot case outside of CPDFSDK_AnnotHandlerMgr.
The work done by CPDFSDK_AnnotHandlerMgr::Annot_OnKeyDown() when the
annot argument is null is better handled by the pageview itself before
calling into the CPDFSDK_AnnotHandlerMgr. This makes the annot handler
manager simpler because it doesn't have to call back into the page
to find an annot. Instead, it performs a single purpose of dispatching
based on the kind of annot it sees.

-- Remove some specific type information from handler mgr, and use
   the virtual interface to ensure uniformity among handler calls.

Change-Id: I4d68cba6035cbb25d003fda96afe3dd91ed2b6fc