Handle null annot case outside of CPDFSDK_AnnotHandlerMgr.
The work done by CPDFSDK_AnnotHandlerMgr::Annot_OnKeyDown() when the
annot argument is null is better handled by the pageview itself before
calling into the CPDFSDK_AnnotHandlerMgr. This makes the annot handler
manager simpler because it doesn't have to call back into the page view
to find an annot. Instead, it performs a single purpose: dispatching
based on the kind of annot it sees.

-- Remove some stray includes.
-- Use incomplete types in .cpp where sufficient.
-- Consistent naming of NewAnnotForXFA() (vs NewXFAAnnot()).
-- Make XFAPageIfNotBackedByPDFPage() a method of the page view.
-- Remove some specific subclass information from handler mgr, and use
   the virtual interface to ensure uniformity among handler calls,
   despite introducing some notreached stubs.

Change-Id: I4d68cba6035cbb25d003fda96afe3dd91ed2b6fc