Re-instate #undef MAKEWORD in fdrm/fx_crypt_aes.cpp.
Follow-up to https://pdfium-review.googlesource.com/c/pdfium/+/85250.

MAKEWORD() is defined several times in this file, so undef after use
is required, but an initial undef seemed like a stray, and I removed
it in https://pdfium-review.googlesource.com/c/pdfium/+/83530. Alas,
it is also defined by windows.

Restore the undef with an explanation, and rename from FMAKEWORD()
back to MAKEWORD(), lest someone down the road decide that removing
the 'F' would make it so much better looking.

Change-Id: Ib58288d412a50dc8d2a1f1776efe0b0659e60b00