Remove unused method CPWL_Edit::SetCharSpace()
In turn, CPWL_EditImpl::SetCharSpace() becomes unused.
In turn, CPVT_VariableText::SetCharSpace() becomes unused.
In turn, CPVT_VariableText::m_fCharSpace is always 0.0.
In turn, CPVT_VariableText::GetCharSpace() is 0.0 and removed.
In turn, CPWL_EditImpl::GetCharSpace() is 0.0 and removed.
In turn, the fCharSpace arg to GetWordWidth() is 0.0 and removed.
In turn, kSetCharacterSpacingOperator becomes unused.

Change-Id: I6e66e6c4a4d7a714005caa363908f4e9ba2f40b4