Make CFFL_InteractiveFormFiller implement IPWL_SystemHandler.
The layering is supposed to be fpdfsdk -> formfiller -> pwl,
however we have some skipping with direct fpdfsdk <--> pwl
dependence. Remove some of this this layer skipping at a cost
of adding interactive form filler methods that proxy upwards
(though they can validate and change arguments to keep the layers
more separate).

This leaves three main PWL usages in fpdfdsk to be reworked:
-- CPWL_Dash, which would have to move to fxcrt/.
-- CPWL_EditImpl, which would become fpdfdoc/CPTV_EditImpl
   (and some sort of drawing callbacks).
-- CPWL_Wnd::IsShiftKeyDown() and friends.

Change-Id: Ia1968cdc53aea91fbbd6b3a68a2cd236ff74fdae