Add non-virtual pre-finalizer entry points.
PDFium's use of virtual pre-finalizers was an artifact of converting
virtual dtors mostly as-is into finalizers. It turns out that cppgc
wasn't supporting this properly, so give it a non-virtual entry point
that can then do the dispatch through the ordinary C++ vtable as
intended.

A subsequent CL can improve the situation by only having the
sub-classes register prefinalizers.

Bug: chromium:1252634
Change-Id: Ibf38de79119d0030fa2f914a5948e2fc4f592430