Fix span calculation in CPDF_DIB::GetScanline()
Ideally, the span size would match the calculated src_pitch_value, but
some spans are actually larger than we think. In cases when there is a
decoder or a cached bitmap, rely on the bounds being correctly set from
those spans, rather than recalculating them based on the current notion
of pitch.

Change-Id: I557880b610af75f8bc68592d462b05cd8be964f2