Enable suppressing test results depending on Skia options.
Adds an extra column in testing/SUPPRESSIONS with possible values of
"*", "noskia", "skia" or "skiapaths" to indicate whether to suppress
the result of a test file depending on the Skia rendering options.

Also modifies the test suppressor to check the Skia build configs
against this new column's value while running tests.

Bug: pdfium:11
Change-Id: I007bfc6be9e497340b8004a84a408977f1a34beb