Add a pixel test for indexing into an object stream.
Entries in a cross-reference stream can reference compressed objects in
an object stream, with an index of the object within the object stream.
Generate an XRef dictionary by hand to use this PDF feature and
demonstrate PDFium not correctly handling the index. The test results
are suppressed because PDFium does not currently render as expected.

The objects in the object stream do not actually need to be compressed.

Bug: pdfium:1733
Change-Id: I986c5fa3768a6aaf4b859bf4d87f4f23379e4bd6