Stop parsing earlier in CPDF_Parser::LoadCrossRefV5() on error.
The object numbers in the /XRef object are in sequencial order. So if a
bad object number has been detected, all subsequent object numbers are
also bad. Thus it is better to stop parsing and just call it quits.

Also rename `objnum` to `obj_num` to be consistent with other variables
in the method.

Change-Id: Ic2f3342c692ecc3d39999bd8175ed68898e9c037