Improve looping behavior in CPDF_Parser::LoadCrossRefV5().
1) Stop parsing earlier in CPDF_Parser::LoadCrossRefV5() on error.

The object numbers in the /XRef object are in sequential order. So if a
bad object number has been detected, all subsequent object numbers are
also bad. Thus it is better to stop parsing and just call it quits.

2) Continue parsing when an invalid archive object number has been
   detected.

There is no reason to exit the method entirely when this happens. Be
consistent with most other error handlers and continue.

Also rename `objnum` to `obj_num` to be consistent with other variables
in the method.

Change-Id: Ic2f3342c692ecc3d39999bd8175ed68898e9c037