Take object index into account for compressed objects.
For a compressed object, its entry in the xref table uses the third
column to store the index into the object stream that stores the
compressed object. Since CPDF_ObjectStream now stores its object info in
a vector, simply index into it to look up the compressed object's offset
in the object stream.

This clears the ambiguity when there are multiple objects in an object
stream with the same object number.

Bug: pdfium:1733
Change-Id: I8a9f52160a3cca6c6c4322f746f9cc599499c21e