Add and fix a CPDF_NameTree unit test for malformed /Limits.
When a /Limits array in a name tree is out of order, CPDF_NameTree tries
to correct that by swapping the 2 array elements. If the /Limits array
is also malformed and has fewer than 2 elements, using
CPDF_Array::SetNewAt() to perform the correction will cause an out of
bounds condition and reach a NOTREACHED(). Fix this by using
InsertNewAt() instead. Add a unit test to demonstrate this issue.

Also document the difference between SetNewAt() and InsertNewAt() in
cpdf_array.h.

Bug: chromium:1263055
Change-Id: Ic7a5ff94d8a1e2612a74daf8d2d9dc25ad733cec