Add and fix a CPDF_NameTree unit test for malformed /Limits.
When a /Limits array in a name tree is out of order, CPDF_NameTree tries
to correct that by using GetNodeLimitsAndSanitize() to swap the 2 array
elements. If the /Limits array is also malformed and has fewer than 2
elements, using CPDF_Array::SetNewAt() to perform the correction will
cause an out of bounds condition and reach a NOTREACHED(). Add a unit
test to demonstrate this issue.

Fix this by not using GetNodeLimitsAndSanitize() anymore. Instead, just
delete the bad /Limits array and reconstruct it.

Bug: chromium:1263055
Change-Id: Ib3828f05078879026fcd2ded5311b6231a97bf7b