Rewrite pixel test bug_665467.
Test bug_665467.in cannot be rendered by macOS Preview and Acrobat due
to invalid "Differences" and "Widths" entry. This CL rewrites the pixel
test so that it uses the font and valid "Differences" and "Widths"
entries which was included in the bug (crbug.com/pdfium/665467).

This CL also updated the expectations for the updated pixel test.

Bug: chromium:665467
Change-Id: I1c646a0ea2c94cacf3472b24ced1c5f58cd959b5