Use consistent base for FGAS_FONTUSB::wBitField sentinel.
We use the value '999' in the table in fgas_fontutils.cpp to indicate
no wBitField value, so expressing it as 0x03e7 in cfgas_fontmgr.cpp
seems deliberately confusing. To be fair, it was originally written
as 999, but changed by us to be more useful on a bit-by-bit basis.
Part of the confusion is likely that wBitField isn't a bitfield
at all, so hex values aren't useful.

Change-Id: I6db5f3cedc0a5320aacc9ea2a532999269238c74