Avoid N^2 copies in CXFA_FMDotAccessorExpression::ToJavaScript().
The local tempExp1 will be repeatedly copied into its caller's
tempExp1 getting larger each time as the recursion unwinds. The
cost now is we may have to generate a string twice, but we do
so into a single buffer.

The referenced bug seems recurse until we hit the 5000 frame
limit in the depth manager.

We'll probably need a follow-up bug to look at the other temporaries.

Bug: chromium:1274018
Change-Id: Ieafcc1e3e1aab412cbcb447f5bcb87d43ab364a2