PreMultiply small text bitmap when Skia is enabled.
On macOS, bitmaps are created with alpha format by default. When a
text is rendered relatively small, it uses DrawNormalTextHelper() to do
gamma adjusting, which results in a bitmap that's not premultiplied.

To avoid crashing the premultiplied check when Skia is enabled, this CL
premultiplies such text bitmaps after gamma adjusting is done.

Bug: pdfium:1757
Change-Id: I36fd500b42ac018f91060450ccd00747e77d7d6e