PreMultiply small text bitmap when Skia is enabled.
On macOS, when bitmap is created with alpha channel by default. When a
text is rendered relatively small, it uses DrawNormalTextHelper() to
render the glyph, which resulted in a bitmap that's not premultiplied.

To avoid crashing the premultiplied check when Skia is enabled,
premultiply such text bitmap before the check happens.

Bug: pdfium:1757
Change-Id: I36fd500b42ac018f91060450ccd00747e77d7d6e