Add a pixel test bug_1355.in.
For corpus test 3bigpreivew.pdf, some Japanese characters are shifted
to the lower left of their expected locations while rendering with
Skia/SkiaPaths on Linux platform.

This CL adds a pixel test of the characters which were rendered
incorrectly and suppress this test when Skia/SkiaPaths is enabled on
Linux. Due to bug_1355.in using the same font as bug_1402.in, this test
also needs to be suppressed on Windows as well.

Bug: pdfium:1355
Change-Id: If228d6e87e6ae4d5e9342732c0c92216e76bde05