Fix a issue of Type3 font rendering.
When a Type3 font's glyph is painted as a bitmap image onto the canvas,
painting it in mask format will result in its rectangle box being
painted with 1 color in fill mode instead of reflecting the glyph's
original bitmap.

Set the SkPaint to use false `isAlphaMask` so that later
SkCanvas::drawImageRect() can paint the glyph's bitmap corretly.

Bug: pdfium:1746,pdfium:1747,pdfium:1749
Change-Id: I36d54d282b7723b5900a494dc987490d5cb5c13a