Update image_transformer_other.in and suppress it for SkiaPaths and AGG.
Acrobat Reader reports error when rendering ImMonoPalette in
image_transformer_other.in since there is not need to use the filter to
decode the stream.

This CL removes the extra "/Filter" entry and updates the expectation
of this test to align with Acrobat's rendering result, in which the
diamond shape on the bottom left is rendered in  two-tone. Also,
suppress this test for AGG and SkiaPaths until crbug.com/pdfium/1750 is
fixed.

Bug: pdfium:1750
Change-Id: I91ec7ca5b56aa181558fe04b3ab6d5da2e4f51c7