Handle building palette from a color space with a hival of 0.
When building palette from an indexed color space, if the color space
has a hival value of 0, its lookup table will only contain 1 color. In
this case, a default color of black (0xFF000000) should also be used for
specifying the range of such color space and affect the palette color.

This fix makes the AGG and SkiaPaths rendering results of
image_transformer_other.in align with Acrobat's.

Bug: pdfium:1750
Change-Id: Iec631ea990d9514932eeca7bc092d3077581e740